Phase 12 : ORM (formulaire pour modifier une visite)

+++ src/Controller/admin/AdminVoyagesController.php
@@ -2,7 +2,9 @@
 
 namespace App\Controller\admin;
 
+use App\Form\VisiteType;
 use App\Repository\VisiteRepository;
+use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -19,10 +21,18 @@ class AdminVoyagesController extends AbstractController {
     
     /**
      * 
-     * @param VisiteRepository $repository
+     * @var EntityManagerInterface
      */
-    public function __construct(VisiteRepository $repository) {
+    private $entityManager;
+    
+    /**
+     * 
+     * @param VisiteRepository $repository
+     * @param EntityManagerInterface $entityManager
+     */
+    public function __construct(VisiteRepository $repository, EntityManagerInterface $entityManager) {
         $this->repository = $repository;
+        $this->entityManager = $entityManager;
     }
     
     #[Route('/admin', name: 'admin.voyages')]
@@ -33,10 +43,34 @@ class AdminVoyagesController extends AbstractController {
         ]);
     }
     
+    #[Route('/admin/edit/{id}', name: 'admin.voyage.edit')]
+    public function edit(int $id, Request $request): Response{
+        $visite = $this->repository->find($id);
+        if (!$visite) {
+            return $this->redirectToRoute('admin.voyages');
+        }
+        
+        $formVisite = $this->createForm(VisiteType::class, $visite);
+        $formVisite->handleRequest($request);
+        
+        if ($formVisite->isSubmitted() && $formVisite->isValid()) {
+            $this->entityManager->persist($visite);
+            $this->entityManager->flush();
+            return $this->redirectToRoute('admin.voyages');
+        }
+        
+        return $this->render("admin/admin.voyage.edit.html.twig", [
+            'visite' => $visite,
+            'formvisite' => $formVisite->createView()
+        ]);
+    }
+    
     #[Route('/admin/suppr/{id}', name: 'admin.voyage.suppr')]
     public function suppr(int $id): Response{
         $visite = $this->repository->find($id);
-        $this->repository->remove($visite);
+        if ($visite) {
+            $this->repository->remove($visite);
+        }
         return $this->redirectToRoute('admin.voyages');
     }
 }
